+++ src/utility/WikiModificationFileInputFormat.java
@@ -2,17 +2,17 @@ package utility;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.RecordReader;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.InputSplit;
 
-public class WikiModificationFileInputFormat extends FileInputFormat<LongWritable, Text> {
+public class WikiModificationFileInputFormat extends FileInputFormat<IntWritable, Text> {
 
 	@Override
-	public RecordReader<LongWritable, Text> createRecordReader(InputSplit split, 
+	public RecordReader<IntWritable, Text> createRecordReader(InputSplit split, 
 			TaskAttemptContext context) throws IOException, InterruptedException {
 		return new RevisionEntryRecordReader();
 	}
